Module5-task2 (#3)
* Добавление наставника в readme

* Merge pull request #2 from Faretos/module1-task1

* Нужно больше функций

* Merge remote-tracking branch 'upstream/master' into module2-task1

* небольшое исправление

* ДЗ Больше деталей модуль 4

* Модуляция

* Небольшое исправление

*  Функции возвращаются

* Merge branch 'master' into module5-task2
