+++ index.html
@@ -235,6 +235,6 @@
     </template>
 
     <script src="/js/functions.js" defer></script>
-    <script src="/js/main.js" defer></script>
+    <script src="/js/main.js" type="module" defer></script>
 </body>
 </html>
